refactor: migrate CLI formatting primitives to TypeScript
