+++ .gitignore
@@ -58,6 +58,8 @@ report/
 .turbo/
 *.cast
 # Generated/temp files
+cli/event-copy.js
+cli/message-formatter-utils.js
 test-metadata-manual.sh
 test-isolated-fix.js
 lib/agent-cli-provider/
