+++ cli/event-copy.ts
@@ -1,12 +1,12 @@
 const EVENT_COPY = {
   IMPLEMENTATION_READY: 'Implementation ready',
   PR_CREATED: 'Pull request created',
-};
+} as const;
 
-function formatMergeStatus(merged) {
+function formatMergeStatus(merged: unknown): string | null {
   if (merged === true || merged === 'true') return 'merged';
   if (merged === false || merged === 'false') return 'auto-merge pending approval';
   return null;
 }
 
-module.exports = { EVENT_COPY, formatMergeStatus };
+export = { EVENT_COPY, formatMergeStatus };
