+++ tests/helpers/runtime-typescript-package-assertions.js
@@ -1,4 +1,6 @@
 const RUNTIME_TYPESCRIPT_OUTPUTS = Object.freeze([
+  'cli/event-copy.js',
+  'cli/message-formatter-utils.js',
   'src/guidance-topics.js',
   'src/input-helpers.js',
   'src/ledger-sequence.js',
